+++ NEWS.md
@@ -1,3 +1,19 @@
+# healthyR.data 1.1.1
+
+## Breaking Changes
+1. Fix #111 - Requires R version 4.1.0 or higher.
+
+## New Function
+1. Fix #102 - Add function `is_valid_url()`
+
+## Minor Fixes and Improvements
+1. Fix #101 - Fixed `fetch_cms_data()` to handle all types of data presented, not 
+just API data. Now appropriately handles both API and non-API data such as, .csv,
+vnd.ms-excel and .zip files.
+2. Fix #103 - Added `is_valid_url()` call to `fetch_` functions.
+3. Fix #107 - Add parameter of `.limit` to `fetch_provider_data()` to limit the number
+of records returned via the API. Default is 500 and 0 will return all records.
+
 # healthyR.data 1.1.0
 
 ## Breaking Changes